Add IAA model
Moves 7600A fields out of the accounts table and into its own model.
Required some refactoring of views and seed script to reflect the new
schema. Also adds dashboard for IAAs and association specs for accounts.
We also add an acronym inflection for "IAA" so that Rails understands it
correctly.

This intentionally leaves a bunch of views and controllers broken. I
will be revamping the schema over the next few commits and then
installing administrate as an initial mechanism for managing the data.