Add status lookup tables for accounts, IAAs, and integrations
Creates a set of three lookup tables to allow for flexible management of
enumerated status values for Accounts, IAAGTCs, IAAOrders, and
Integrations. IAAGTCs and IAAOrders share the same set of statuses. Each
lookup table record consists of a unique name and a unique integer
order, the latter of which is used to determine the display order of
statuses in dropdowns.

I originally attempted to "be fancy" and have all of the `belongs_to`
associations named `:status` and use additional parameters to identify
the class_name and foreign key, but this made things messier than they
needed to be and in the worst case we should be able to eventually
override the attribute labels if we need to. Therefore everything here
ended up pretty vanilla.

This also includes some minor changes to the flash message overrides in
the IntegrationContact and AccountContact controllers for (slightly)
more consistent capitalization. It also includes a slight reordering of
the dashboards in Administrate and the seeding of default statuses in
the seed script.