Add partner-facing name to status lookup tables (#21)
Add an additional attribute to the three status lookup tables called
`partner_name` that is intended for use on partner-facing screens. This
allows us to have multiple internal statuses that correspond to a single
partner-facing status. Also modify the getter method for that attribute
to fall back to the `name` attribute if no `partner_name` is set.