Integrate the login.gov design system
* Replace the USDS with the login.gov design system
* Tweak some existing views to better leverage USDS classes
* Add FlashesHelper to manage the different keys that Devise uses and
  make them compatible with the USDS alerts

Note that we are not modifying the administrate styling at this time. I
did attempt to integrate the login design system into the administrate
stylesheet to add the two banners and things started breaking, so if we
want to adjust the administrate views to leverage standard login layouts
we can do that separately.